requirements.txt
flask
numpy
requests
beautifulsoup4
python-whois
scikit-learn